re-install packaje.json
