+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-portfolio",
+  "name": "EdsonSanchez",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
